refactor: move internal RPC methods to `rpc.ts`
