+++ src/internal/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './events';
+export * from './rpc';
 export * from './types';
